+++ R/AccumulationExceedingThreshold.R
@@ -113,7 +113,7 @@ CST_AccumulationExceedingThreshold <- function(data, threshold, op = '>', diff =
                                           na.rm = na.rm, ncores = ncores)
   data$data <- total
   data$dims <- dim(total)
-  data$coords[[time_dim]] <- NULL
+  data$coords[[time_dim]] <- 1 : length(data$dims[[time_dim]])
 
   if (!is.null(Dates)) {
     if (!is.null(start) && !is.null(end)) {
@@ -128,10 +128,10 @@ CST_AccumulationExceedingThreshold <- function(data, threshold, op = '>', diff =
       # Create time_bounds
       time_bounds <- NULL
       time_bounds$start <- ClimProjDiags::Subset(x = Dates, along = time_dim, 
-                                                 indices = 1, drop = 'selected')
+                                                 indices = 1, drop = FALSE)
       time_bounds$end <- ClimProjDiags::Subset(x = Dates, along = time_dim, 
                                                indices = dim(Dates)[time_dim], 
-                                               drop = 'selected')
+                                               drop = FALSE)
 
       # Add Dates in attrs
       data$attrs$Dates <- time_bounds$start
@@ -203,6 +203,7 @@ CST_AccumulationExceedingThreshold <- function(data, threshold, op = '>', diff =
 #'GDD <- AccumulationExceedingThreshold(data, threshold = 0, start = list(1, 4),
 #'                                      end = list(31, 10))
 #'@import multiApply
+#'@importFrom stats setNames
 #'@export
 AccumulationExceedingThreshold <- function(data, threshold, op = '>', diff = FALSE,
                                            dates = NULL, start = NULL, end = NULL,
@@ -436,6 +437,7 @@ AccumulationExceedingThreshold <- function(data, threshold, op = '>', diff = FAL
                      ncores = ncores)$output1
     }
   }
+  dim(total) <- c(dim(total), setNames(1, time_dim))
   return(total)
 }