+++ R/SelectPeriodOnData.R
@@ -121,7 +121,7 @@ SelectPeriodOnData <- function(data, dates, start, end,
                ncores = ncores)$output1
   # when 29Feb is included the length of the output changes:
   regular <- Apply(list(res), target_dims = time_dim,
-                   fun = sum, ncores = ncores)$output1
+                   fun = function(...) {sum(...)}, ncores = ncores)$output1
   dims <- dim(data)
   dims[names(dims) == time_dim] <- max(regular)
   if (any(regular != max(regular))) {